# services/app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Generación de PDFs",
    "version": "1.0.0",
    "description": "API para el sistema de diseño y generación de documentos PDF"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
  "paths": {
...
    "/templates": {
      "get": {
        "tags": ["Plantillas"],
        "summary": "Obtener plantillas",
        "description": "Lista todas las plantillas a las que el usuario tiene acceso",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de plantillas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Template" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Plantillas"],
        "summary": "Crear plantilla",
        "description": "Crea una nueva plantilla con su versión inicial",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TemplateCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plantilla creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/Template" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/templates/{template_id}": {
      "parameters": [
        {
          "name": "template_id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" }
        }
      ],
      "get": {
        "tags": ["Plantillas"],
        "summary": "Obtener plantilla",
        "description": "Obtiene una plantilla específica",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Plantilla encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/Template" }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Plantillas"],
        "summary": "Actualizar plantilla",
        "description": "Actualiza una plantilla existente",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TemplateUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plantilla actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/Template" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Plantillas"],
        "summary": "Eliminar plantilla",
        "description": "Elimina una plantilla y sus recursos",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Plantilla eliminada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/templates/{template_id}/versions": {
      "parameters": [
        {
          "name": "template_id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" }
        }
      ],
      "get": {
        "tags": ["Versiones"],
        "summary": "Obtener versiones",
        "description": "Obtiene todas las versiones de una plantilla",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de versiones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TemplateVersion" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Versiones"],
        "summary": "Crear versión",
        "description": "Crea una nueva versión de la plantilla",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": { "$ref": "#/components/schemas/TemplateContent" }
                },
                "required": ["content"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Versión creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/TemplateVersion" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/templates/{template_id}/resources": {
      "parameters": [
        {
          "name": "template_id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" }
        }
      ],
      "get": {
        "tags": ["Recursos"],
        "summary": "Obtener recursos",
        "description": "Obtiene todos los recursos de una plantilla",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de recursos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TemplateResource" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Recursos"],
        "summary": "Subir recurso",
        "description": "Sube un nuevo recurso para la plantilla",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "resource_type": {
                    "type": "string",
                    "enum": ["image", "logo", "other"]
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recurso subido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/TemplateResource" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Template": {
        "type": "object",
        "properties": {
          "template_id": { "type": "integer" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "category_id": { "type": "integer" },
          "is_active": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "created_by": { "type": "integer" },
          "updated_by": { "type": "integer" }
        }
      },
      "TemplateCreate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "category_id": { "type": "integer" },
          "content": { "$ref": "#/components/schemas/TemplateContent" },
          "resources": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ResourceUpload" }
          },
          "permissions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PermissionAssignment" }
          }
        },
        "required": ["name", "category_id"]
      },
      "TemplateUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "category_id": { "type": "integer" },
          "is_active": { "type": "boolean" }
        }
      },
      "TemplateContent": {
        "type": "object",
        "properties": {
          "header": { "type": "string" },
          "body": { "type": "string" },
          "footer": { "type": "string" }
        },
        "required": ["header", "body", "footer"]
      },
      "TemplateVersion": {
        "type": "object",
        "properties": {
          "version_id": { "type": "integer" },
          "template_id": { "type": "integer" },
          "version_number": { "type": "integer" },
          "content": { "$ref": "#/components/schemas/TemplateContent" },
          "created_at": { "type": "string", "format": "date-time" },
          "created_by": { "type": "integer" },
          "is_active": { "type": "boolean" }
        }
      },
      "TemplateResource": {
        "type": "object",
        "properties": {
          "resource_id": { "type": "integer" },
          "resource_type": { "type": "string" },
          "original_name": { "type": "string" },
          "url": { "type": "string" }
        }
      },
      "ResourceUpload": {
        "type": "object",
        "properties": {
          "file_data": { "type": "string", "format": "binary" },
          "resource_type": { "type": "string" },
          "original_name": { "type": "string" }
        },
        "required": ["file_data", "original_name"]
      },
      "PermissionAssignment": {
        "type": "object",
        "properties": {
          "role_id": { "type": "integer" },
          "can_view": { "type": "boolean" },
          "can_edit": { "type": "boolean" },
          "can_generate": { "type": "boolean" }
        },
        "required": ["role_id"]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}